+++ src/main/java/org/depromeet/sambad/moring/meeting/question/domain/MeetingQuestionStatus.java
@@ -1,0 +1,5 @@
+package org.depromeet.sambad.moring.meeting.question.domain;
+
+public enum MeetingQuestionStatus {
+	ACTIVE, INACTIVE, NOT_STARTED;
+}
